Reformate code
